Modularise bashrc and implement a caching mechanism.
Split bashrc into {function,alias,complete}.bash, each of which contains
various checks and magic to miniise environment pollution. *.bash files
are sourced unsing the newly defined $RC_DIR variable, enabling their
use as dotfiles in a traditional enviroment, or inside a dotfiles
directory as this repo (assuming the bashrc is symlinked to the
directory.

After sourcing, the resultant environmnet is cached into
respective files held on tmpfs, so that successive shell start-ups are
much faster, as they can avoid all the checks.

.{rc,ps1,alias,function,complete} aliases have also been defined to allow
the reloading of individual components (and consequent bypass of cache).

TODO:
Implement an hash check on the files so as to bypass the cache when they
have been modified.