Shell Config update:
Enforce a stricter naming convention for global shell variables should
be all capitals, and if they describe directories they should have a
'_D' suffix (not nothing or '_DIR', etc.).
Keep USE_COLOR, and then use it in 'alias.sh' to define ls and grep
aliases. Removes hackery needed otherwise to re-alias them, and allows
the moving of all alias definitions to 'alias.sh'.
Add a 'cd' wrapper function to complement my '_cdd' completion, allowing
the actual command to recognise the '....' shorthand as well as the
completion.
Use ``sed /proc/mounts'' as a check for /dev/shm, not a path detection
as it is a POSIX standard value if it exists.
Add an 'autostart.sh' functionality, where each line in the file is read
by bashrc (as cmd args), then run with (as much as possible of):
    sudo -b -u ${USER} nice ionice ${cmd} ${args} >/dev/null 2>&1
so as to background it and provide a minimum of interference.