Fix bash completion functions.
Fix both the 'cd' function (from 'function.sh') and the relevent
completion function '_cdd', so that there is correct quoting with no
arguments, and the completion works regardless of position or options
(so it can be used with pushd as well.
Fix the '_wrap_alias' function so it works properly with 'ls' aliases.
The problem was only COMP_CWORD and COMP_WORDS were updated, but
COMP_POINT and COMP_LINE needed to be as well.